Ta peganu fogo meu

+++ server/oladoutor/src/main/java/br/com/model/PatientDrug.java
@@ -25,7 +25,7 @@ public class PatientDrug implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "id", updatable = false, nullable = false)
-	private Long id;
+	private Integer id;
 	@Version
 	@Column(name = "version")
 	private int version;
@@ -47,11 +47,11 @@ public class PatientDrug implements Serializable {
 	@Temporal(TemporalType.TIMESTAMP)
 	private Date time;
 
-	public Long getId() {
+	public Integer getId() {
 		return this.id;
 	}
 
-	public void setId(final Long id) {
+	public void setId(final Integer id) {
 		this.id = id;
 	}
